rename status.h to priority.h
